requirements.txt
bson==0.5.2
certifi==2018.1.18
chardet==3.0.4
click==6.7
configparser==3.5.0
docopt==0.6.2
dominate==2.3.1
ExifRead==2.1.2
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Script==2.0.6
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Pillow==5.0.0
pymongo==3.6.1
pytz==2018.3
requests==2.18.4
six==1.11.0
Werkzeug==0.14.1
WTForms==2.1
gunicorn==19.7.1
